shoppet/wsgi.py
"""
WSGI config for shoppet project.
"""

import os
from django.core.wsgi import get_wsgi_application
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'shoppet.settings')

application = get_wsgi_application()